# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary List all conversations
 */
export const ListOpenaiConversationsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiConversationsResponse = zod.array(
  ListOpenaiConversationsResponseItem,
);

/**
 * @summary Create a new conversation
 */
export const CreateOpenaiConversationBody = zod.object({
  title: zod.string(),
});

/**
 * @summary Get conversation with messages
 */
export const GetOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOpenaiConversationResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  createdAt: zod.date(),
  messages: zod.array(
    zod.object({
      id: zod.number(),
      conversationId: zod.number(),
      role: zod.string(),
      content: zod.string(),
      createdAt: zod.date(),
    }),
  ),
});

/**
 * @summary Delete a conversation
 */
export const DeleteOpenaiConversationParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List messages in a conversation
 */
export const ListOpenaiMessagesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListOpenaiMessagesResponseItem = zod.object({
  id: zod.number(),
  conversationId: zod.number(),
  role: zod.string(),
  content: zod.string(),
  createdAt: zod.date(),
});
export const ListOpenaiMessagesResponse = zod.array(
  ListOpenaiMessagesResponseItem,
);

/**
 * @summary Send a text message and receive a streaming text response
 */
export const SendOpenaiMessageParams = zod.object({
  id: zod.coerce.number(),
});

export const SendOpenaiMessageBody = zod.object({
  content: zod.string(),
  seenBefore: zod.boolean().optional(),
});

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});